requirements.txt
requests
python-dotenv
PyPDF2
PyGithub
langchain
langchain-openai
langchain-community
langchain-core

redis
redis-py-cluster
